Modularized the files to separate concerns
